Correction ejercicio002
